chore: restructured the pharaoh aignment heuristic and set noise value from 0 to 1 (double instead of int)

+++ KhetApi/Models/Move/Move.cs
@@ -23,7 +23,7 @@ public sealed class UndoState
 
 public class SearchResult
 {
-    public int Score;
+    public double Score;
     public Move BestMove;
 }
